Add files via upload
1.	Introduction

1.1	Background
Green spaces are areas of untouched or maintained wilderness as well as parks inside of cities (urban parks). There is empirical evidence that green spaces promote increased mental health and general well being. They also make exercise and activities more available to lower income individuals. People who move to greener spaces have been shown to have sustained improvements in mood.

1.2	Problem
Minneapolis, MN is a city with some of the highest numbers of green spaces in the United States. If someone were to move to Minneapolis and wanted the highest number of parks within a walking distance, which one of the 11 communities should they move into? 
2.	Data

2.1	Data Sources
The data used in this project will be from Foursquare and Wikipedia. The list of the 11 communities in Minneapolis was found on Wikipedia and a CSV document of each community and the coordinates of their geographical center (found on google maps) was created. Foursquare was used for parks and lakes location data. Lakes were considered green spaces as they are usually surrounded by nature to protect the watershed. 
2.2	Data Cleaning and Wrangling
The CSV file did not have to cleaned or wrangled since it was created by me to match my data needs. It was imported into Watson Studio to be used as a pandas data frame. The parks and lakes location data were requested from Foursquare and made into 2 pandas data frames. It was cleaned to only include columns with the venue name and any location data. Duplicates locations were also removed.